Handle wizard submission, remove WizardStore
